Calls API endpoint 'DataStructure'

Calls API endpoint 'DataStructure'

Usage

mt_data_structure(id)

Arguments

id

is a Dataset Series identifier.

Value

a list with content of 'DataStructure' response.

Examples

mt_data_structure("DOT")
#> $Structure
#> $Structure$`@xmlns:xsd`
#> [1] "http://www.w3.org/2001/XMLSchema"
#> 
#> $Structure$`@xmlns:xsi`
#> [1] "http://www.w3.org/2001/XMLSchema-instance"
#> 
#> $Structure$`@xmlns`
#> [1] "http://www.SDMX.org/resources/SDMXML/schemas/v2_0/message"
#> 
#> $Structure$`@xsi:schemaLocation`
#> [1] "http://www.SDMX.org/resources/SDMXML/schemas/v2_0/structure https://registry.sdmx.org/schemas/v2_0/SDMXStructure.xsd http://www.SDMX.org/resources/SDMXML/schemas/v2_0/message https://registry.sdmx.org/schemas/v2_0/SDMXMessage.xsd"
#> 
#> $Structure$Header
#> $Structure$Header$ID
#> [1] "9b0dbfe2-6176-43e2-a9db-cac88fcfbbbc"
#> 
#> $Structure$Header$Test
#> [1] "false"
#> 
#> $Structure$Header$Prepared
#> [1] "2023-11-08T13:45:31"
#> 
#> $Structure$Header$Sender
#> $Structure$Header$Sender$`@id`
#> [1] "1C0"
#> 
#> $Structure$Header$Sender$Name
#> $Structure$Header$Sender$Name$`@xml:lang`
#> [1] "en"
#> 
#> $Structure$Header$Sender$Name$`#text`
#> [1] "IMF"
#> 
#> 
#> $Structure$Header$Sender$Contact
#> $Structure$Header$Sender$Contact$URI
#> [1] "http://www.imf.org"
#> 
#> $Structure$Header$Sender$Contact$Telephone
#> [1] "+ 1 (202) 623-6220"
#> 
#> 
#> 
#> $Structure$Header$Receiver
#> $Structure$Header$Receiver$`@id`
#> [1] "ZZZ"
#> 
#> 
#> 
#> $Structure$CodeLists
#> $Structure$CodeLists$CodeList
#>                       @id @agencyID @version @isFinal
#> 1            CL_UNIT_MULT       IMF      1.0     true
#> 2                 CL_FREQ       IMF      1.0     true
#> 3             CL_AREA_DOT       IMF      1.0     true
#> 4        CL_INDICATOR_DOT       IMF      1.0     true
#> 5 CL_COUNTERPART_AREA_DOT       IMF      1.0     true
#> 6          CL_TIME_FORMAT       IMF      1.0     true
#>                                                        @xmlns
#> 1 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/structure
#> 2 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/structure
#> 3 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/structure
#> 4 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/structure
#> 5 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/structure
#> 6 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/structure
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Code
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    0, 2, 3, 6, 9, 12, N15, N14, N13, N12, N11, N10, N9, N8, N7, N6, N5, N4, N3, N2, N1, 1, 4, 5, 7, 8, 10, 11, 13, 14, 15, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, Units, Hundreds, Thousands, Millions, Billions, Trillions, Quadrillionths, Hundred Trillionths, Ten Trillionths, Trillionths, Hundred Billionths, Ten Billionths, Billionths, Hundred Millionths, Ten Millionths, Millionths, Hundred Thousandths, Ten Thousandths, Thousandths, Hundredths, Tenths, Tens, Ten Thousands, Hundred Thousands, Ten Millions, Hundred Millions, Ten Billions, Hundred Billions, Ten Trillions, Hundred Trillions, Quadrillions
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     A, B, Q, M, D, W, en, en, en, en, en, en, Annual, Bi-annual, Quarterly, Monthly, Daily, Weekly
#> 3 AF, F19, AL, DZ, AS, AO, AI, AG, AR, AM, AW, S19, AU, AT, AZ, BS, BH, BD, BB, BY, BE, R1, BZ, BJ, BM, BT, BO, BA, BW, BR, BN, BG, BF, BI, CV, KH, CM, CA, CF, TD, CL, HK, MO, CN, CO, KM, CD, CG, CR, CI, _X, HR, CU, CW, CY, CZ, CSH, DK, DJ, DM, DO, DE2, EC, EG, SV, GQ, ER, EE, SZ, ET, E19, FK, FO, FJ, FI, FR, PF, GA, GM, GE, DE, GH, GI, GR, GL, GD, GU, GT, GN, GW, GY, HT, VA, HN, HU, IS, IN, ID, IR, IQ, IE, IL, IT, JM, JP, JO, KZ, KE, KI, KP, KR, XK, KW, KG, LA, LV, LB, LS, LR, LY, LT, LU, MG, MW, MY, 1C_554, MV, ML, MT, MH, MR, MU, MX, FM, F979, MD, MN, ME, MS, MA, MZ, MM, NA, NR, NP, AN, NL, NC, NZ, NI, NE, NG, MK, 1C_958, NO, OM, PK, PW, PA, PG, PY, PE, PH, PL, PT, QA, RO, RU, RW, WS, SM, ST, SA, SN, CS, RS, SC, SL, SG, SX, SK, SI, SB, SO, ZA, 1C_198, SS, ES, XR99, LK, KN, LC, VC, SD, SR, SE, CH, SY, TW, TJ, TZ, TH, TL, TG, TO, TT, TN, TR, TM, TV, UG, UA, AE, GB, US, UY, SUH, UZ, VU, VE, VN, PS, A109, 1C_473, 1C_459, YE, YUC, ZM, ZW, XR29, F1, R14, XS25, XR43, 1C_903, 1C_080, 1C_092, U2, E1, B0, F97, XS5, 1C_440, X88, F6, A10, W00, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, Afghanistan, Africa not allocated, Albania, Algeria, American Samoa, Angola, Anguilla, Antigua and Barbuda, Argentina, Armenia, Aruba, Asia not allocated, Australia, Austria, Azerbaijan, Bahamas, Bahrain, Bangladesh, Barbados, Belarus, Belgium, Belgo-Luxembourg Economic Union, Belize, Benin, Bermuda, Bhutan, Bolivia, Bosnia and Herzegovina, Botswana, Brazil, Brunei Darussalam, Bulgaria, Burkina Faso, Burundi, Cabo Verde, Cambodia, Cameroon, Canada, Central African Republic, Chad, Chile, Hong Kong, China, Macao, China, Colombia, Comoros, Congo, the Democratic Republic of the, Congo, Costa Rica, Cote d`Ivoire, Not allocated/unspecified, Croatia, Cuba, Curacao, Cyprus, Czech Republic, Former Czechoslovakia, Denmark, Djibouti, Dominica, Dominican Republic, East Germany, Ecuador, Egypt, El Salvador, Equatorial Guinea, Eritrea, Estonia, Eswatini, Ethiopia, Europe not allocated, Falkland Islands (Malvinas), Faroe Islands, Fiji, Finland, France, French Polynesia, Gabon, Gambia, Georgia, Germany, Ghana, Gibraltar, Greece, Greenland, Grenada, Guam, Guatemala, Guinea, Guinea-Bissau, Guyana, Haiti, Holy See (Vatican City State), Honduras, Hungary, Iceland, India, Indonesia, Iran, Islamic Republic of, Iraq, Ireland, Israel, Italy, Jamaica, Japan, Jordan, Kazakhstan, Kenya, Kiribati, Korea, Democratic People`s Republic of, Korea, Republic of, Kosovo, Republic of, Kuwait, Kyrgyzstan, Lao People`s Democratic Republic, Latvia, Lebanon, Lesotho, Liberia, Libya, Lithuania, Luxembourg, Madagascar, Malawi, Malaysia, Malaysia, West, Maldives, Mali, Malta, Marshall islands, Mauritania, Mauritius, Mexico, Micronesia, Federated States of, Middle East and Central Asia not specified, Moldova, Republic of, Mongolia, Montenegro, Montserrat, Morocco, Mozambique, Myanmar, Namibia, Nauru, Nepal, Netherlands Antilles, Netherlands, New Caledonia, New Zealand, Nicaragua, Niger, Nigeria, North Macedonia, Republic of , North Vietnam, Norway, Oman, Pakistan, Palau, Panama, Papua New Guinea, Paraguay, Peru, Philippines, Poland, Portugal, Qatar, Romania, Russian Federation, Rwanda, Samoa, San Marino, Sao Tome and Principe, Saudi Arabia, Senegal, Serbia and Montenegro, Serbia, Seychelles, Sierra Leone, Singapore, Sint Maarten (Dutch part), Slovakia, Slovenia, Solomon Islands, Somalia, South Africa, South African Common Customs Area (SACCA), South Sudan, Spain, Special Categories and Economic Zones, Sri Lanka, Saint Kitts and Nevis, Saint Lucia, Saint Vincent and the Grenadines, Sudan, Suriname, Sweden, Switzerland, Syrian Arab Republic, Taiwan, Province of China, Tajikistan, Tanzania, United Republic of, Thailand, Timor-Leste, Togo, Tonga, Trinidad and Tobago, Tunisia, Türkiye, Turkmenistan, Tuvalu, Uganda, Ukraine, United Arab Emirates, United Kingdom, United States, Uruguay, Former U.S.S.R., Uzbekistan, Vanuatu, Venezuela, Bolivarian Republic, Viet Nam, Palestine, State of, Western Hemisphere not allocated, Yemen Arab Rep., Yemen, P.D. Rep., Yemen, Former Yugoslavia, Zambia, Zimbabwe, Advanced Economies (IMF), Africa, Community of Independent States (CIS), Developing Asia (IMF), Emerging and Developing Countries, Emerging and Developing Europe, Export earnings: fuel, Export earnings: nonfuel, Euro Area (Member States and Institutions of the Euro Area) changing composition, Europe, EU (Member States and Institutions of the European Union) changing composition, Middle East, Middle East and Central Asia, Middle East, North Africa, Afghanistan, and Pakistan, Other Countries n.i.e. (IMF), Sub-Saharan Africa, Western Hemisphere, All Countries, excluding the IO
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 TXG_FOB_USD, TMG_CIF_USD, TMG_FOB_USD, TBG_USD, en, en, en, en, Goods, Value of Exports, Free on board (FOB), US Dollars, Goods, Value of Imports, Cost, Insurance, Freight (CIF), US Dollars, Goods, Value of Imports, Free on board (FOB), US Dollars, Goods, Value of Trade Balance, US Dollars
#> 5 AF, F19, AL, DZ, AS, AO, AI, AG, AR, AM, AW, S19, AU, AT, AZ, BS, BH, BD, BB, BY, BE, R1, BZ, BJ, BM, BT, BO, BA, BW, BR, BN, BG, BF, BI, CV, KH, CM, CA, CF, TD, CL, HK, MO, CN, CO, KM, CD, CG, CR, CI, _X, HR, CU, CW, CY, CZ, CSH, DK, DJ, DM, DO, DE2, EC, EG, SV, GQ, ER, EE, SZ, ET, E19, FK, FO, FJ, FI, FR, PF, GA, GM, GE, DE, GH, GI, GR, GL, GD, GU, GT, GN, GW, GY, HT, VA, HN, HU, IS, IN, ID, IR, IQ, IE, IL, IT, JM, JP, JO, KZ, KE, KI, KP, KR, XK, KW, KG, LA, LV, LB, LS, LR, LY, LT, LU, MG, MW, MY, 1C_554, MV, ML, MT, MH, MR, MU, MX, FM, F979, MD, MN, ME, MS, MA, MZ, MM, NA, NR, NP, AN, NL, NC, NZ, NI, NE, NG, MK, 1C_958, NO, OM, PK, PW, PA, PG, PY, PE, PH, PL, PT, QA, RO, RU, RW, WS, SM, ST, SA, SN, CS, RS, SC, SL, SG, SX, SK, SI, SB, SO, ZA, 1C_198, SS, ES, XR99, LK, KN, LC, VC, SD, SR, SE, CH, SY, TW, TJ, TZ, TH, TL, TG, TO, TT, TN, TR, TM, TV, UG, UA, AE, GB, US, UY, SUH, UZ, VU, VE, VN, PS, A109, 1C_473, 1C_459, YE, YUC, ZM, ZW, XR29, F1, R14, XS25, XR43, 1C_903, 1C_080, 1C_092, U2, E1, B0, F97, XS5, 1C_440, X88, F6, A10, W00, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, en, Afghanistan, Africa not allocated, Albania, Algeria, American Samoa, Angola, Anguilla, Antigua and Barbuda, Argentina, Armenia, Aruba, Asia not allocated, Australia, Austria, Azerbaijan, Bahamas, Bahrain, Bangladesh, Barbados, Belarus, Belgium, Belgo-Luxembourg Economic Union, Belize, Benin, Bermuda, Bhutan, Bolivia, Bosnia and Herzegovina, Botswana, Brazil, Brunei Darussalam, Bulgaria, Burkina Faso, Burundi, Cabo Verde, Cambodia, Cameroon, Canada, Central African Republic, Chad, Chile, Hong Kong, China, Macao, China, Colombia, Comoros, Congo, the Democratic Republic of the, Congo, Costa Rica, Cote d`Ivoire, Not allocated/unspecified, Croatia, Cuba, Curacao, Cyprus, Czech Republic, Former Czechoslovakia, Denmark, Djibouti, Dominica, Dominican Republic, East Germany, Ecuador, Egypt, El Salvador, Equatorial Guinea, Eritrea, Estonia, Eswatini, Ethiopia, Europe not allocated, Falkland Islands (Malvinas), Faroe Islands, Fiji, Finland, France, French Polynesia, Gabon, Gambia, Georgia, Germany, Ghana, Gibraltar, Greece, Greenland, Grenada, Guam, Guatemala, Guinea, Guinea-Bissau, Guyana, Haiti, Holy See (Vatican City State), Honduras, Hungary, Iceland, India, Indonesia, Iran, Islamic Republic of, Iraq, Ireland, Israel, Italy, Jamaica, Japan, Jordan, Kazakhstan, Kenya, Kiribati, Korea, Democratic People`s Republic of, Korea, Republic of, Kosovo, Republic of, Kuwait, Kyrgyzstan, Lao People`s Democratic Republic, Latvia, Lebanon, Lesotho, Liberia, Libya, Lithuania, Luxembourg, Madagascar, Malawi, Malaysia, Malaysia, West, Maldives, Mali, Malta, Marshall islands, Mauritania, Mauritius, Mexico, Micronesia, Federated States of, Middle East and Central Asia not specified, Moldova, Republic of, Mongolia, Montenegro, Montserrat, Morocco, Mozambique, Myanmar, Namibia, Nauru, Nepal, Netherlands Antilles, Netherlands, New Caledonia, New Zealand, Nicaragua, Niger, Nigeria, North Macedonia, Republic of , North Vietnam, Norway, Oman, Pakistan, Palau, Panama, Papua New Guinea, Paraguay, Peru, Philippines, Poland, Portugal, Qatar, Romania, Russian Federation, Rwanda, Samoa, San Marino, Sao Tome and Principe, Saudi Arabia, Senegal, Serbia and Montenegro, Serbia, Seychelles, Sierra Leone, Singapore, Sint Maarten (Dutch part), Slovakia, Slovenia, Solomon Islands, Somalia, South Africa, South African Common Customs Area (SACCA), South Sudan, Spain, Special Categories and Economic Zones, Sri Lanka, Saint Kitts and Nevis, Saint Lucia, Saint Vincent and the Grenadines, Sudan, Suriname, Sweden, Switzerland, Syrian Arab Republic, Taiwan, Province of China, Tajikistan, Tanzania, United Republic of, Thailand, Timor-Leste, Togo, Tonga, Trinidad and Tobago, Tunisia, Türkiye, Turkmenistan, Tuvalu, Uganda, Ukraine, United Arab Emirates, United Kingdom, United States, Uruguay, Former U.S.S.R., Uzbekistan, Vanuatu, Venezuela, Bolivarian Republic, Viet Nam, Palestine, State of, Western Hemisphere not allocated, Yemen Arab Rep., Yemen, P.D. Rep., Yemen, Former Yugoslavia, Zambia, Zimbabwe, Advanced Economies (IMF), Africa, Community of Independent States (CIS), Developing Asia (IMF), Emerging and Developing Countries, Emerging and Developing Europe, Export earnings: fuel, Export earnings: nonfuel, Euro Area (Member States and Institutions of the Euro Area) changing composition, Europe, EU (Member States and Institutions of the European Union) changing composition, Middle East, Middle East and Central Asia, Middle East, North Africa, Afghanistan, and Pakistan, Other Countries n.i.e. (IMF), Sub-Saharan Africa, Western Hemisphere, All Countries, excluding the IO
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         P1Y, P6M, P3M, P1M, P7D, P1D, en, en, en, en, en, en, Annual, Bi-annual, Quarterly, Monthly, Weekly, Daily
#>   Name.@xml:lang                     Name.#text Description.@xml:lang
#> 1             en                          Scale                  <NA>
#> 2             en                      Frequency                    en
#> 3             en             Geographical Areas                  <NA>
#> 4             en                      Indicator                  <NA>
#> 5             en Counterpart Geographical Areas                  <NA>
#> 6             en                    Time format                    en
#>                 Description.#text
#> 1                            <NA>
#> 2                       Frequency
#> 3                            <NA>
#> 4                            <NA>
#> 5                            <NA>
#> 6 Time formats based on ISO 8601.
#> 
#> 
#> $Structure$Concepts
#> $Structure$Concepts$ConceptScheme
#> $Structure$Concepts$ConceptScheme$`@id`
#> [1] "DOT"
#> 
#> $Structure$Concepts$ConceptScheme$`@agencyID`
#> [1] "IMF"
#> 
#> $Structure$Concepts$ConceptScheme$`@version`
#> [1] "1.0"
#> 
#> $Structure$Concepts$ConceptScheme$`@isFinal`
#> [1] "true"
#> 
#> $Structure$Concepts$ConceptScheme$`@xmlns`
#> [1] "http://www.SDMX.org/resources/SDMXML/schemas/v2_0/structure"
#> 
#> $Structure$Concepts$ConceptScheme$Name
#> $Structure$Concepts$ConceptScheme$Name$`@xml:lang`
#> [1] "en"
#> 
#> $Structure$Concepts$ConceptScheme$Name$`#text`
#> [1] "Direction of Trade Statistics (DOTS)"
#> 
#> 
#> $Structure$Concepts$ConceptScheme$Concept
#>                @id @agencyID @version Name.@xml:lang
#> 1        OBS_VALUE       IMF      1.0             en
#> 2        UNIT_MULT       IMF      1.0             en
#> 3      TIME_FORMAT       IMF      1.0             en
#> 4             FREQ       IMF      1.0             en
#> 5         REF_AREA       IMF      1.0             en
#> 6        INDICATOR       IMF      1.0             en
#> 7 COUNTERPART_AREA       IMF      1.0             en
#> 8      TIME_PERIOD       IMF      1.0             en
#> 9       OBS_STATUS       IMF      1.0             en
#>                                   Name.#text TextFormat.@textType
#> 1                                      Value               Double
#> 2                                      Scale                 <NA>
#> 3                                Time format                 <NA>
#> 4                                  Frequency                 <NA>
#> 5                             Reference Area                 <NA>
#> 6                                  Indicator                 <NA>
#> 7                 Counterpart Reference Area                 <NA>
#> 8                                       Date             DateTime
#> 9 Observation Status (incl. Confidentiality)               String
#>   Description.@xml:lang               Description.#text
#> 1                  <NA>                            <NA>
#> 2                  <NA>                            <NA>
#> 3                    en Time formats based on ISO 8601.
#> 4                  <NA>                            <NA>
#> 5                  <NA>                            <NA>
#> 6                  <NA>                            <NA>
#> 7                  <NA>                            <NA>
#> 8                  <NA>                            <NA>
#> 9                  <NA>                            <NA>
#> 
#> 
#> 
#> $Structure$KeyFamilies
#> $Structure$KeyFamilies$KeyFamily
#> $Structure$KeyFamilies$KeyFamily$`@id`
#> [1] "DOT"
#> 
#> $Structure$KeyFamilies$KeyFamily$`@agencyID`
#> [1] "IMF"
#> 
#> $Structure$KeyFamilies$KeyFamily$`@version`
#> [1] "1.0"
#> 
#> $Structure$KeyFamilies$KeyFamily$`@isFinal`
#> [1] "true"
#> 
#> $Structure$KeyFamilies$KeyFamily$`@xmlns`
#> [1] "http://www.SDMX.org/resources/SDMXML/schemas/v2_0/structure"
#> 
#> $Structure$KeyFamilies$KeyFamily$Name
#> $Structure$KeyFamilies$KeyFamily$Name$`@xml:lang`
#> [1] "en"
#> 
#> $Structure$KeyFamilies$KeyFamily$Name$`#text`
#> [1] "Direction of Trade Statistics (DOTS)"
#> 
#> 
#> $Structure$KeyFamilies$KeyFamily$Components
#> $Structure$KeyFamilies$KeyFamily$Components$Dimension
#>        @conceptRef @conceptVersion @conceptSchemeRef @conceptSchemeAgency
#> 1             FREQ             1.0               DOT                  IMF
#> 2         REF_AREA             1.0               DOT                  IMF
#> 3        INDICATOR             1.0               DOT                  IMF
#> 4 COUNTERPART_AREA             1.0               DOT                  IMF
#>                 @codelist @codelistVersion @codelistAgency
#> 1                 CL_FREQ              1.0             IMF
#> 2             CL_AREA_DOT              1.0             IMF
#> 3        CL_INDICATOR_DOT              1.0             IMF
#> 4 CL_COUNTERPART_AREA_DOT              1.0             IMF
#>   @isFrequencyDimension
#> 1                  true
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$TimeDimension
#> $Structure$KeyFamilies$KeyFamily$Components$TimeDimension$`@conceptRef`
#> [1] "TIME_PERIOD"
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$TimeDimension$`@conceptVersion`
#> [1] "1.0"
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$TimeDimension$`@conceptSchemeRef`
#> [1] "DOT"
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$TimeDimension$`@conceptSchemeAgency`
#> [1] "IMF"
#> 
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$PrimaryMeasure
#> $Structure$KeyFamilies$KeyFamily$Components$PrimaryMeasure$`@conceptRef`
#> [1] "OBS_VALUE"
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$PrimaryMeasure$`@conceptVersion`
#> [1] "1.0"
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$PrimaryMeasure$`@conceptSchemeRef`
#> [1] "DOT"
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$PrimaryMeasure$`@conceptSchemeAgency`
#> [1] "IMF"
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$PrimaryMeasure$TextFormat
#> $Structure$KeyFamilies$KeyFamily$Components$PrimaryMeasure$TextFormat$`@textType`
#> [1] "Double"
#> 
#> 
#> 
#> $Structure$KeyFamilies$KeyFamily$Components$Attribute
#>   @conceptRef @conceptVersion @conceptSchemeRef @conceptSchemeAgency
#> 1   UNIT_MULT             1.0               DOT                  IMF
#> 2  OBS_STATUS             1.0               DOT                  IMF
#> 3 TIME_FORMAT             1.0               DOT                  IMF
#>        @codelist @codelistVersion @codelistAgency @attachmentLevel
#> 1   CL_UNIT_MULT              1.0             IMF           Series
#> 2           <NA>             <NA>            <NA>      Observation
#> 3 CL_TIME_FORMAT              1.0             IMF           Series
#>   @assignmentStatus @isTimeFormat TextFormat.@textType
#> 1         Mandatory          <NA>                 <NA>
#> 2       Conditional          <NA>               String
#> 3         Mandatory          true                 <NA>
#> 
#> 
#> $Structure$KeyFamilies$KeyFamily$Annotations
#> $Structure$KeyFamilies$KeyFamily$Annotations$Annotation
#>                                                     @xmlns     AnnotationTitle
#> 1 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/common  Latest Update Date
#> 2 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/common                Name
#> 3 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/common   Temporal Coverage
#> 4 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/common Geographic Coverage
#> 5 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/common         Methodology
#> 6 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/common          Definition
#> 7 http://www.SDMX.org/resources/SDMXML/schemas/v2_0/common                Code
#>   AnnotationText.@xml:lang
#> 1                       en
#> 2                       en
#> 3                       en
#> 4                       en
#> 5                       en
#> 6                       en
#> 7                       en
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                AnnotationText.#text
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        10/24/2023
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Direction of Trade Statistics (DOTS)
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Monthly and quarterly data are available starting 1960. Annual data are available starting 1947.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         DOTS covers all IMF member states, some non-member countries, the world, and major areas.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Guide to Direction of Trade Statistics, 1993. See Documents tab.
#> 6 The <B>Direction of Trade Statistics (DOTS)</B> presents the value of merchandise exports and imports disaggregated according to a country's primary trading partners. Area and world aggregates are included in the display of trade flows between major areas of the world. Reported data is supplemented by estimates whenever such data is not available or current. Imports are reported on a cost, insurance and freight (CIF) basis and exports are reported on a free on board (FOB) basis, with the exception of a few countries for which imports are also available FOB. Time series data includes estimates derived from reports of partner countries for non-reporting and slow-reporting countries.
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               DOT
#> 
#> 
#> 
#> 
#>